astro-prediction-analyzer/requirements.txt
youtube-transcript-api==0.6.1
google-auth-oauthlib==1.2.0
google-api-python-client==2.107.0
google-cloud-translate==3.11.0
openai>=1.0.0
python-dotenv==1.0.0
pandas==2.0.3
pydantic==2.5.0
requests==2.31.0
tenacity==8.2.3
tqdm==4.66.1
langdetect==1.0.9
googleapis-common-protos==1.63.0
